Drop ARMv6 support
It is with a heavy heart that I do this but we are reaching the limits of the Pi 1 and Zero v1.x CPU (they aren't powerful enough for volume normalization) and librespot will soon be merging a rewrite that is incompatible with ARMv6 (it will not compile).

The upside to this is that since we are now purely relying on vanilla Debian Stable for our libs (except librespot ofc) that should mean wider compatability with other Debian based armhf distros and it opens up the possibility of including the pulseaudio backend in the builds.