Make librespot abort on panic and not hang
The downside to this is that we won't get tracebacks on panics, but I think that's Ok for the most part, as error handling in librespot has gotten much better in recent releases.

The upsides are that systemd should beable to restart the Raspotify service now if/when librespot panics and since we're no longer using any of the traceback info we can strip the binary making the .deb about half the size. (not that it was huge in the 1st place)