Add channel Up/Down Mixing
Channel Up Mixing is done with channel duplication for the Right and Left Surround Channels and channel combination for the Center and LFE (Sub) Channels without any effects or high/low pass filtering.

It is not true surround sound.

Channel Down Mixing is done with channel combination for Mono.

Some Devices do not map their channels correctly.

In that case manually editing of /etc/asound.conf may be necessary.