installer: fix recent Debian and Ubuntu versions
and add support for riscv64.

Ubuntu 24.04 Noble and Debian 13 Trixie made the 64-bit `time_t` transition, and the breaking API change due to related `time64` syscalls on 32-bit systems is expressed with a t64 suffix in library package names. `libasound2` has hence been renamed to `libasound2t64`.

On 64-bit systems, where `time_t` values were 64-bit before, the transition is a noop, and hence related packages "provide" the old ones. Hence `apt install libasound2` on 64-bit systems still works, being satisfied by the actually existing `libasound2t64`. But on 32-bit systems this command fails.

However, `libasound2` on Ubuntu is also "provided" by an alternative package `liboss4-salsa-asound2`, and the same is available on Debian Forky/testing. With multiple alternative "providers" for an only virtual package defined as dependency, APT denies to do the choice, but requires one of the providers to be chosen directly. This additionally breaks `apt install libasound2` on 64-bit systems on any Ubuntu or Debian version with `libasound2t64` and `liboss4-salsa-asound2` both available.

The solution applied is was to remove `libasound2` from explicit dependencies. The correct `libasound2` variant is a dependency of `alsa-utils` on all Debian and Ubuntu versions, and both always share the same version string. Since `alsa-utils` is explicitly installed as dependency, and its version checked, `libasound2` is hence not required.

An alternative solution would be to remove all these dependencies from the installer. They are actually all not required since the `raspotify` package itself declares everything already, for APT to require and pull the required dependencies automatically. In case of an `apt-get` failure, on top of its own error message, the general information about which architectures are support could still be printed, and a cleanup/removal of APT key+list done.

Signed-off-by: MichaIng <micha@dietpi.com>