installer: remove manual dependency installation
This is not needed, since the DEB package declares all dependencies properly and APT installs them automatically.

Aside of reducing maintenance efforts, a benefit on admin end is that dependencies which were installed for Raspotify only, can be `apt autopurge`d.

Also, the installer uses now `apt-get` in any case. `apt` is not intended to be used in scripts, and throws a warning if not attached to an interactive terminal:
```
WARNING: apt does not have a stable CLI interface. Use with caution in scripts.
```

And finally, add the `-f`/`--fail` flag to the `curl` call. That way, in case of HTTP errors, a one-line error message is printed to STDERR, hence no HTML error-document is stored to the key file. To also prevent an existing key from being emptied on error, the `-o` option is used to store the successful download, instead of piping STDOUT.

Signed-off-by: MichaIng <micha@dietpi.com>